+++ src/editor/YamlEditor.vue
@@ -2,11 +2,11 @@
 import { computed, useTemplateRef } from "vue";
 import { Codemirror } from "vue-codemirror";
 import type { Extension } from "@codemirror/state";
-import type { ViewUpdate } from "@codemirror/view";
+import { EditorView, type ViewUpdate } from "@codemirror/view";
 import { monokai } from "@fsegurai/codemirror-theme-monokai";
 import { computedAsync, useDark, useLocalStorage } from "@vueuse/core";
 import { PhFileArrowUp } from "@phosphor-icons/vue";
-import { diagnosticCount, linter } from "@codemirror/lint";
+import { diagnosticCount, linter, lintGutter } from "@codemirror/lint";
 import { indentWrapExtension } from "./indentWrap";
 import ActionButton from "@/components/ActionButton.vue";
 import FileDropArea from "@/components/FileDropArea.vue";
@@ -38,11 +38,23 @@ const validationExtension = computedAsync(async () => {
   const YamlValidator = yamlValidator.default;
   const validator = new YamlValidator(props.schema);
 
-  return linter((view) => {
-    const text = view.state.doc.toString();
-    const errors = validator.validate(text);
-    return errors.map((error) => ({ ...error, severity: "error" }));
+  const linterExtension = linter(
+    (view) => {
+      const text = view.state.doc.toString();
+      const errors = validator.validate(text);
+      return errors.map((error) => ({ ...error, severity: "error" }));
+    },
+    { autoPanel: true },
+  );
+
+  // The linter sets a light background, regardless of dark mode. Undo that.
+  const linterTheme = EditorView.theme({
+    ".cm-panel.cm-panel-lint ul [aria-selected]": {
+      backgroundColor: "inherit",
+    },
   });
+
+  return [linterExtension, linterTheme, lintGutter()];
 });
 
 /** Lazy-loaded YAML language support extension */