+++ src/page/AppHeader.vue
@@ -20,35 +20,72 @@ const { isPending } = useSpin();
 
 const isHome = computed(() => route.path == "/");
 const isAuthenticating = computed(() => isPending("jwt"));
+
+/** Replacement for the RouterLink `active-class` prop, because it only checks route nesting, not the path string. */
+const isActiveClass = (path: string) =>
+  route.path.startsWith(path) ? "text-sborange-600" : "";
 </script>
 
 <template>
   <header class="mb-2 shadow-sm bg-white dark:bg-zinc-800">
+    <!-- Main row -->
     <div
-      class="container pt-4 pb-2 flex justify-between items-baseline-last gap-4"
+      class="container pt-4 pb-2 flex max-sm:flex-col flex-wrap sm:items-baseline-last gap-4 gap-y-2"
     >
       <!-- Logo -->
       <component
         :is="isHome ? 'h1' : 'div'"
-        class="text-4xl min-w-max"
+        class="text-4xl min-w-max mr-8 max-sm:mb-2"
         :class="large ? 'h-24' : 'h-16'"
       >
-        <router-link to="/" class="text-current">
-          <MinkLogo :large="large" />
+        <router-link to="/" class="text-current" :title="$t('home')">
+          <MinkLogo :large />
         </router-link>
       </component>
 
-      <SpinIndicator v-if="isAuthenticating" />
-
-      <!-- Navigation -->
+      <!-- Main navigation -->
       <div
-        class="flex max-md:flex-col gap-x-4 items-baseline justify-end text-end text-nowrap"
+        class="flex sm:max-lg:flex-col gap-x-4 items-start text-nowrap text-lg font-medium"
       >
+        <router-link
+          to="/"
+          class="no-underline hover:underline"
+          exact-active-class="text-sborange-600"
+        >
+          {{ $t("home") }}
+        </router-link>
+
+        <router-link
+          to="/library"
+          class="no-underline hover:underline"
+          :class="isActiveClass('/library')"
+        >
+          {{ $t("library") }}
+        </router-link>
+
+        <router-link
+          to="/tools"
+          class="no-underline hover:underline"
+          :class="isActiveClass('/tools')"
+        >
+          {{ $t("tools") }}
+        </router-link>
+      </div>
+
+      <!-- Spacing spinner in the middle, always taking space, but visible only while auth is loading -->
+      <SpinIndicator
+        class="max-sm:hidden grow mx-auto"
+        :class="{ invisible: !isAuthenticating }"
+      />
+
+      <div class="flex sm:max-lg:flex-col gap-x-4 text-nowrap">
+        <!-- User link if logged in -->
         <template v-if="isAuthenticated">
           <router-link
             v-if="canUserWrite"
             to="/user"
             class="no-underline hover:underline"
+            :class="isActiveClass('/user')"
           >
             <PhUser class="inline mb-0.5 mr-1" />{{ userName }}
           </router-link>
@@ -57,6 +94,7 @@ const isAuthenticating = computed(() => isPending("jwt"));
           </a>
         </template>
 
+        <!-- Help link -->
         <a
           :href="$t('docs.url')"
           target="_blank"
@@ -69,6 +107,7 @@ const isAuthenticating = computed(() => isPending("jwt"));
       </div>
     </div>
 
+    <!-- Admin mode banner row -->
     <AdminModeBanner />
   </header>
 </template>